+++ src/main/java/cow/farm/repo/CowHerdService.java
@@ -9,7 +9,7 @@ public class CowHerdService {
 
 	QuantityCheckingList<Cow> cowList = new QuantityCheckingList<>();
 
-	public Cow makeCow(String name, Cow.Breed breed, float weight, Cow.Color color) {
+	public Cow makeCow(String name, Cow.Breed breed, short weight, Cow.Color color) {
 		Cow cow = new Cow(name, breed, weight, color);
 		cowList.add(cow);
 		return cow;
@@ -23,7 +23,7 @@ public class CowHerdService {
 		return cowList.getQuantityOfEachAttribute(new CowToBreedTransformation());
 	}
 
-	public Map<Float, Integer> getHerdQuantityByWeight() {
+	public Map<Short, Integer> getHerdQuantityByWeight() {
 		return cowList.getQuantityOfEachAttribute(new CowToWeightTransformation());
 	}
